Create SearchCrypto
Replace 'your_token_here' with your actual Telegram bot token. This code defines a Telegram bot that responds to /start and /search commands. When a user sends a /search command followed by a query, the bot will search for information related to that query in the cryptosphere (in the search_cryptosphere function).

You'll need to implement the search_cryptosphere function to search for information in the cryptosphere. This could involve using APIs from cryptocurrency websites, databases, or other sources to fetch relevant information based on the query.